changed file name of compare_embedding to create_database
